backend/requirements.txt
# FastAPI 核心（使用更新版本，支持 Python 3.13）
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0

# 数据库
# 选项1: 使用 psycopg3（推荐，支持 Python 3.13）
psycopg[binary]>=3.2.2
# 选项2: 如果上面不行，可以尝试更新的 psycopg2-binary
# psycopg2-binary>=2.9.10

# SQLAlchemy 需要使用更新版本以支持 Python 3.13
sqlalchemy>=2.0.36
alembic>=1.13.0

# 图片处理（使用更新的稳定版本，有预编译的 Windows wheel）
Pillow>=10.2.0

# S3 客户端（用于 RustFS/S3 兼容存储）
aioboto3==12.3.0
boto3==1.34.0

# 工具类
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# CORS (内置在 FastAPI 中，无需单独安装)

